Add files via upload
In this project, a device classification architecture is proposed that can distinguish IoT devices from non-IoT devices using a feature set of IoT-specific traffic.
The outcome is a Deep Learning Model which distinguishes IoT and non-IoT devices with higher accuracy and as well as classify IoT devices with the required accuracy.	The results produced by this architecture will be helpful for the operator of the smart home environment to manage IoT devices by monitoring them and restrict the malicious activities through them.